relocate the tests and set bits
