+++ src/ubigint/std.rs
@@ -69,3 +69,80 @@ impl<const N_BITS: usize> UBigIntImpl<N_BITS> {
         }
     }
 }
+
+#[cfg(test)]
+mod test {
+    use rand_chacha::ChaCha20Rng;
+    use rand::{Rng, SeedableRng};
+    use bitcoin_script::script;
+    use crate::treepp::{execute_script, pushable};
+    use crate::ubigint::UBigIntImpl;
+
+    #[test]
+    fn test_zip() {
+        const N_BITS: usize = 1500;
+        const N_U30_LIMBS: usize = 50;
+
+        let mut prng = ChaCha20Rng::seed_from_u64(0);
+
+        for _ in 0..50 {
+            let mut v = vec![];
+            for _ in 0..N_U30_LIMBS {
+                v.push(prng.gen::<i32>());
+            }
+            for _ in 0..N_U30_LIMBS {
+                v.push(prng.gen::<i32>());
+            }
+
+            let mut expected = vec![];
+            for i in 0..N_U30_LIMBS {
+                expected.push(v[i]);
+                expected.push(v[N_U30_LIMBS + i]);
+            }
+
+            let script = script! {
+                for i in 0..(N_U30_LIMBS * 2) {
+                    { v[i as usize] }
+                }
+                { UBigIntImpl::<N_BITS>::zip(1, 0) }
+                for i in 0..(N_U30_LIMBS * 2) as u32 {
+                    { expected[N_U30_LIMBS * 2 - 1 - (i as usize)] }
+                    OP_EQUALVERIFY
+                }
+                OP_PUSHNUM_1
+            };
+            let exec_result = execute_script(script);
+            assert!(exec_result.success);
+        }
+
+        for _ in 0..50 {
+            let mut v = vec![];
+            for _ in 0..N_U30_LIMBS {
+                v.push(prng.gen::<i32>());
+            }
+            for _ in 0..N_U30_LIMBS {
+                v.push(prng.gen::<i32>());
+            }
+
+            let mut expected = vec![];
+            for i in 0..N_U30_LIMBS {
+                expected.push(v[N_U30_LIMBS + i]);
+                expected.push(v[i]);
+            }
+
+            let script = script! {
+                for i in 0..N_U30_LIMBS * 2 {
+                    { v[i] }
+                }
+                { UBigIntImpl::<N_BITS>::zip(0, 1) }
+                for i in 0..N_U30_LIMBS * 2 {
+                    { expected[N_U30_LIMBS * 2 - 1 - (i)] }
+                    OP_EQUALVERIFY
+                }
+                OP_PUSHNUM_1
+            };
+            let exec_result = execute_script(script);
+            assert!(exec_result.success);
+        }
+    }
+}
